Create README.md
 /*
  usage
  
  Compute range:
  vkeys: search key column
  vfields: header row
  
  Data range:
  vkeys2: key column at the data range
  range: data range. first row is the lookup header row
  */

VLOOKUP2D
Vertical lookup. Uses specified key field in a 2D data field and prepares the data in the specified columns.  

Sample Usage
=VLOOKUP2D(A:A, A:D, F1:Z1,F:Z )

Syntax
=VLOOKUP2D(key_column, report_header_range, data_ data_key_column, data_range)

Assumes that the unique values in the selected key column. 

report_key_column - 1st column in the report with specified values to search for. For example, Name: Ali, Veli, Selami

report_header_range - Column headers of specified values to match across the target. For example: Age, Height

data_key_column - the key column in the data. it is not necessarily the first column. 

data_range - The range to consider for the search. For example: Name, Age, Height, City, State

The data_key_column column in the range is searched for the key values specified in report_key_column.

Exact , case sensitive, and unique matches are reported. Null fields are returned for no matches.